# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API.",
        "title": "REST API for recipes backend.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
        "/ingredients": {
            "get": {
                "description": "get all ingredients",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all ingredients",
                "operationId": "allingredients",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Ingredient"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add ingredient",
                "produces": [
                    "application/json"
                ],
                "summary": "Add an ingredient",
                "operationId": "addingredient",
                "parameters": [
                    {
                        "description": "Ingredient",
                        "name": "ingredient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.IngredientWithoutID"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Ingredient"
                        }
                    }
                }
            }
        },
        "/ingredients/generate": {
            "post": {
                "description": "Generate ingredients",
                "produces": [
                    "application/json"
                ],
                "summary": "Generate ingredients",
                "operationId": "generateingredients",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Ingredient"
                            }
                        }
                    }
                }
            }
        },
        "/ingredients/{name}": {
            "get": {
                "description": "get ingredient by name",
                "produces": [
                    "application/json"
                ],
                "summary": "Get ingredient by name",
                "operationId": "ingredientbyname",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Ingredient"
                        }
                    }
                }
            }
        },
        "/recipes": {
            "get": {
                "description": "get all recipes",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all recipes",
                "operationId": "allrecipes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Recipe"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "get a recipe by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Add a recipe",
                "operationId": "addrecipe",
                "parameters": [
                    {
                        "description": "Recipe",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Recipe"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.RecipeWithoutID"
                        }
                    }
                }
            }
        },
        "/recipes/generate": {
            "post": {
                "description": "Generate recipes",
                "produces": [
                    "application/json"
                ],
                "summary": "Generate recipe",
                "operationId": "generaterecipe",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Recipe"
                            }
                        }
                    }
                }
            }
        },
        "/recipes/{id}": {
            "get": {
                "description": "get a recipe by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get recipe by ID",
                "operationId": "getrecipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Recipe"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Category": {
            "type": "string",
            "enum": [
                "DRINK",
                "MAIN_COURSE",
                "DESSERT",
                "APPETIZER"
            ],
            "x-enum-varnames": [
                "CategoryDrink",
                "CategoryMainCourse",
                "CategoryDessert",
                "CategoryAppetizer"
            ]
        },
        "model.Ingredient": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.IngredientMeta": {
            "type": "object",
            "properties": {
                "quantity": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "model.IngredientWithoutID": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Recipe": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "category": {
                    "$ref": "#/definitions/model.Category"
                },
                "description": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ingredients_meta": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.IngredientMeta"
                    }
                },
                "name": {
                    "type": "string"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.RecipeWithoutID": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/model.Category"
                },
                "description": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ingredients_meta": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.IngredientMeta"
                    }
                },
                "name": {
                    "type": "string"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}